+++ Realm/RLMDefines.h
@@ -30,13 +30,11 @@
 #define RLM_GENERIC_COLLECTION <RLMObjectType: RLMObject *>
 #define RLM_GENERIC_RETURN <RLMObjectType>
 #define RLMObjectArgument RLMObjectType
-#define RLM_GENERIC_ARRAY(CLASS) RLMArray<CLASS *><CLASS>
 #else
 #define RLM_GENERIC_COLLECTION
 #define RLM_GENERIC_RETURN
 typedef id RLMObjectType;
 typedef RLMObject * RLMObjectArgument;
-#define RLM_GENERIC_ARRAY(CLASS) RLMArray<CLASS>
 #endif
 
 #pragma mark - Nullability